+++ DeMobile/Models/AppModel/m_Customer.cs
@@ -38,6 +38,6 @@ namespace DeMobile.Models.AppModel
         public int api_version { get; set; }
         public string pin { get; set; }
         public DateTime created_time { get; set; }
-        public int app_version { get; set; }
+        public double app_version { get; set; }
     }
 }
